requirements.txt
configobj==5.0.6
loguru==0.5.3
pandas==1.4.0
PyQt5==5.15.6
PyYAML==6.0
requests==2.27.1
tqdm==4.62.3
